backend/requirements.txt
google-auth-oauthlib
google-api-python-client
google-auth
isodate
openai
dotenv
uvicorn==0.23.2